constrastive loss exp introduces.

+++ new_train_and_eval.py
@@ -72,7 +72,7 @@ def train_sage(model, g, feats, optimizer, epoch, logger):
     scaler = torch.cuda.amp.GradScaler()
     optimizer.zero_grad()
 
-    with torch.cuda.amp.autocast():
+    with torch.cuda.amp.autocast(dtype=torch.float16):
         _, logits, loss, _, cb, loss_list3, latent_train, quantized, latents, sample_list_train = model(g, feats, epoch,
                                                                                                         logger)  # g is blocks
     loss = loss.to(device)
